Add support for cross-compilation
Two options have been defined for `nodec` in order to properly define
--dest-os and --dest-arch whose values are actually the same
one can pass down to the node.js configuration.

When those values are defined we also define some other needed flags
to prevent node.js compilation to execute target binaries on the host.
This is needed because binaries are meant for another architecture/os
than the current one.

Applications for this feature are related to the ability to compile
for another device/processor architecture, or even utilize an older
linux box to compile for newer kernels.

I also managed to compile Mac OS X to Linux but in order to do so
one must be able to perform the `mkpeephole` step on the target
os directly or using a virtualized machine.

I think it would be great to also be able to define at runtime or
install time the node.js version and get the tarball at runtime
with the source code one wants.

Note: in order to perform cross-compilation one must be able to
setup a cross-compile toolchain properly (crosstool-ng is advised
to greatly simplify this task).