Add a flag to skip npm install
It may be desirable to skip the npm install process when relying on other form
of automation, such as build systems or CI which may already setup the environment
and install packages for us. These packages will then not be statically put inside
our binary, which contains only the application (no node_modules/).
Meaning that the external system will need to make sure everything is in place
to run the binary correctly (e.g. set the proper environment variables such as NODE_PATH).